Encode Shannon Information

info_encoder converts the i into distinct series based on monotonically occurring breaks.
The cumulative proportionality within each series is converted into Shannon information as "bits".

info_encoder(i, info.only = FALSE, data.only = info.only)

Arguments

i

(numeric) The input vector of differences

info.only

(logical) TRUE returns the information column only

data.only

(logical) TRUE returns columns series and cyl only, and FALSE returns everything

Value

If info.only is TRUE, a numeric vector of Shannon information is returned, otherwise a data frame with columns series, cyl, and info is returned.

Note

info_encoder() requires the input to be the differences of an ordered vector

See also