Signal Processor

signal_processor processes the "signal" of a "break_signal" object.

signal_processor(object, cl_size = 1, ...)

Arguments

object

A "break_signal" S7 object

cl_size

(integer|1) The number of parallel workers to use during processing if greater than one.

...

Arguments for internal use

Value

The modified "break_signal" object with prescribed slots populated from the results of processing the "signal" (see break_signal).

See also