Continuity Creator

continuity is conceptually based on the 'islands & gaps' concept.

continuity(
  data,
  map_fields,
  time_fields,
  timeout = 0,
  boundary_name = "window",
  archipelago = TRUE,
  show.all = FALSE
)

Arguments

data

(object): The source dataset, including all non-sessioning fields desired

map_fields

(string): A comma-separated string literal containing field names that will partition data

time_fields

(string): A comma-separated string literal containing field names to use as "start" and "stop" temporal indices. If only one value is given, that value will be repeated as the "stop" index

timeout

The largest allowable 'gap' in a series of time values before a new 'island' begins: can be a quoted expression that conditionally determines the value.

boundary_name

(string): The name root of the boundary column names (e.g., "episode" -> "episode_start_idx", "episode_end_idx")

archipelago

(logical | TRUE): Should the output include the islands and gaps generated?

show.all

(logical | FALSE): Should the output include all of the columns of the output?

Value

A data.table with columns , ..., timeOut, ISLAND, island_idx, where '...' is empty if show.all is FALSE