Event Vectors Class

event.vectors is an R6 class object that creates a temporally-compared space of "event vectors", each comprised of a 'from' and 'to' temporal marker.

The time between these events is the focus of derivation: properly, the event vector is a complex number that encodes the relationship between the boundaries of the events, thus allowing one to describe this relationship in a concise manner.

Execution Workflow

The initial execution order should look something like the following ...
event.vectors$new(...)$configure(...)$make.evs_universe(...) . The ability to execute the preceding workflow out of order exists, but it is best to adhere to the provided flow the first time around.

Class Member "Space"

$space is a data.table that is populated upon execution of $make.evs_universe():

jk

Values of the "join key"

"crossed" time output

See cross.time()

fr_*, to_*

The temporal ranges of each from/toevent

src.pair

Values of the "from-to" pairings of events given by the contexts provided: FROM:TO

from.coord

String representation of temporal boundaries expressed as concatenated integers

to.coord

String representation of temporal boundaries expressed as concatenated integers

Public fields

space

The object containing the derived set of data (see 'Details')

q_graph

A 'query graph' list containing the event interactions for each level of 'jk' (see 'Details')

evt_graphs

Event graphs for each level of 'jk' (see 'Details') Initialize the Event Vector Class

Active bindings

config

This is an active binding that returns the configuration used to instantiate the class.

Methods

Public methods

event.vectors$new()

event.vectors$configure()

event.vectors$make.evs_universe()

event.vectors$clone()

Method new()

Usage

event.vectors$new(...)

Arguments

...

(not used)

Method configure()

$configure() creates references to the source data to use.

Usage

event.vectors$configure(
  src.defs,
  contexts,
  map.fields = NULL,
  src.mix = "comb",
  exclude.mix = NULL,
  chatty = FALSE
)

Arguments

src.defs

Strings or expressions that define the data to use (must return data.frame, data.table, or list)

contexts

An atomic vector of strings serving as labels for each data source

map.fields

A list of column name vectors for each definition in src.defs providing data for the join key (jk), start (time_start_idx), and end (time_end_idx) references. Provide names for elements in each vector to specify which data source field is mapped to each name (jk, time_start_idx, and time_end_idx).

src.mix

"combination":

The default: generates unique pairs of sources

"reflexive":

Compares each source to itself

"all":

The union of the preceding options

Partial matching is allowed, and generated combinations include the mirror (e.g., "A, B" will have a "B, A" combo pair generated).

exclude.mix

A list of vectors containing each source pair context to exclude (e.g. list(c("A", "C"), c("u", "k"))). evs_exclude.blender can be invoked to create this list more quickly. Combinations are not automatically mirrored as is the case with src.mix.

chatty

(logical | FALSE) Verbosity flag

Method make.evs_universe()

make.evs_universe supplies values to two class fields: q_graph and space, the latter being created from the former.

Additional Information:

Class member $space should have as many rows as the sum of all edge counts for graphs in $q_graph

The graphs in class member $evt_graphs are visIgraph-ready

Parallelism is internally supported via package furrr: the user is responsible for setting the appropriate plan

Usage

event.vectors$make.evs_universe(
  ...,
  time.control = list(-Inf, Inf),
  graph.control = NULL,
  unit = "",
  furrr_opts = furrr::furrr_options(scheduling = Inf, seed = TRUE),
  graph.only = FALSE,
  chatty = FALSE
)

Arguments

...

(dots_list) Logical expression that retain graph edges meeting the conditions

time.control

A 2-element list containing the minimum and maximum values allowed for total temporal span between two events

graph.control

An expression list containing igraph-package calls to manipulate the internally-created graph in the order provided. Use symbol g to generically denote the graph and <<- to update g.

unit

(See cross.time)

furrr_opts

furrr_options defaulted as scheduling = Inf and seed = TRUE: internal globals are also set and will be appended to values provided here

graph.only

(logical | FALSE) TRUE assumes class member $space exists (possibly after external modification) and recreates member $evt_graphs

chatty

(logical | FALSE) Verbosity flag

Returns

Invisibly, the original object augmented with new member space

Method clone()

The objects of this class are cloneable with this method.

Usage

event.vectors$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.